+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## v2.2.0 (in development)
 
-No changes yet.
+### Enhancements
+
+ * `parameters list_all` is a new command that lists all runtime parameters across all virtual hosts
+ * `parameters list_in` is a new command that lists runtime parameters of a given component (type)
+   in a specific virtual host
 
 
 ## v2.1.0 (May 8, 2025)